Refactor: factory 메서드는 utils 에 속하도록 변경
