+++ src/main/java/io/powerrangers/backend/service/CustomOauth2UserService.kt
@@ -7,6 +7,7 @@ import io.powerrangers.backend.entity.User
 import io.powerrangers.backend.exception.AuthTokenException
 import io.powerrangers.backend.exception.ErrorCode
 import io.powerrangers.backend.utils.toUserDetails
+import io.powerrangers.backend.utils.userDetails
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest
